This is Refactor of project: 1- versions of dependencies are updated and security is removed( just for a while) 2- flyway and lombok is added 3- thymleaf is changed and html files are updated and fragment is added

+++ src/main/java/com/amin/service/CountryService.java
@@ -1,13 +1,16 @@
 package com.amin.service;
 
-import com.amin.entity.Country;
+import com.amin.domain.entity.Country;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface CountryService {
     public List<Country> findAll();
-    public Optional<Country> findById(int id);
+
+    public Optional<Country> findById(Long id);
+
     public void save(Country country);
-    public void deleteById(int id);
+
+    public void deleteById(Long id);
 }
